Move gltf plugins
