+++ src/plugins/index.d.ts
@@ -5,3 +5,8 @@ export * from './three/plugins/DebugTilesPlugin';
 
 // common plugins
 export { ImplicitTilingPlugin } from './base/plugins/ImplicitTilingPlugin';
+
+// gltf extensions
+export { GLTFCesiumRTCExtension } from './three/gltf/GLTFCesiumRTCExtension';
+export { GLTFStructuralMetadataExtension } from './three/gltf/GLTFStructuralMetadataExtension';
+export { GLTFMeshFeaturesExtension } from './three/gltf/GLTFMeshFeaturesExtension';
