+++ src/plugins/index.js
@@ -5,3 +5,8 @@ export * from './three/plugins/DebugTilesPlugin.js';
 
 // common plugins
 export { ImplicitTilingPlugin } from './base/plugins/ImplicitTilingPlugin.js';
+
+// gltf extensions
+export { GLTFCesiumRTCExtension } from './three/loaders/gltf/GLTFCesiumRTCExtension.js';
+export { GLTFStructuralMetadataExtension } from './three/loaders/gltf/GLTFStructuralMetadataExtension.js';
+export { GLTFMeshFeaturesExtension } from './three/loaders/gltf/GLTFMeshFeaturesExtension.js';
