+++ src/plugins/three/gltf/GLTFCesiumRTCExtension.d.ts
