+++ src/plugins/three/gltf/GLTFMeshFeaturesExtension.d.ts
