+++ src/plugins/three/gltf/GLTFStructuralMetadataExtension.d.ts
