+++ src/plugins/three/gltf/metadata/classes/ClassProperty.js
