+++ src/plugins/three/gltf/metadata/classes/MeshFeatures.js
