+++ src/plugins/three/gltf/metadata/classes/PropertyAttributeAccessor.js
