+++ src/plugins/three/gltf/metadata/classes/PropertySetAccessor.js
