+++ src/plugins/three/gltf/metadata/classes/PropertyTableAccessor.js
