+++ src/plugins/three/gltf/metadata/classes/PropertyTextureAccessor.js
