+++ src/plugins/three/gltf/metadata/classes/StructuralMetadata.js
