+++ src/plugins/three/gltf/metadata/utilities/ClassPropertyHelpers.js
