+++ src/plugins/three/gltf/metadata/utilities/TexCoordUtilities.js
