+++ src/plugins/three/gltf/metadata/utilities/TextureReadUtility.js
